Refactor and add new util func
